+++ spec/models/group_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Group do
       raw_group = double(:raw_group, :uuid => 'uuid', :description => 'desc', :name => 'gname', :principals => %w(u1 u2))
       group_api = double(:group_api)
       expect(group_api).to receive(:get_group).with('uuid').and_return(raw_group)
-      expect(RBACService).to receive(:call).with(RBACApiClient::GroupApi).and_yield(group_api)
+      expect(RBAC::Service).to receive(:call).with(RBACApiClient::GroupApi).and_yield(group_api)
     end
 
     it 'fetches a group with details from rbac service' do
@@ -23,7 +23,7 @@ RSpec.describe Group do
       raw_list = double(:group_list, :meta => double(:count => 2), :data => raw_groups)
       group_api = double(:group_api)
       expect(group_api).to receive(:list_groups).with(hash_including(:username => 'myname')).and_return(raw_list)
-      expect(RBACService).to receive(:call).with(RBACApiClient::GroupApi).and_yield(group_api)
+      expect(RBAC::Service).to receive(:call).with(RBACApiClient::GroupApi).and_yield(group_api)
     end
 
     it 'list all groups' do
